build: remove the tools folder (#24880)
